+++ packages/react/src/index.ts
@@ -1,7 +1,7 @@
-export {retain, release} from '@remote-ui/rpc';
 export {createRemoteRoot} from '@remote-ui/core';
 export type {RemoteRoot, RemoteReceiver} from '@remote-ui/core';
-export {render} from './render';
+export {render, createRoot} from './render';
+export type {Root} from './render';
 export {createRemoteReactComponent} from './components';
 export {useRemoteSubscription} from './hooks';
 export type {